Update project licenses
The licenses for both the Bootstrap and main project have been updated. The previous GNU Public License was replaced with a shorter, simplified custom 'Akami Solutions Public License'. Additionally, a new MIT license was added for the Bootstrap library. This change will provide clearer guidelines for using and contributing to the project, improving transparency and legal clarity.