Update Docker publishing workflows
The workflows for Docker image publishing have been revised. The changes are centered around using GitHub Container Registry instead of Docker Hub and supporting additional branches. The workflows are now triggered when changes are pushed to the `development` and `production` branches and the Docker images are stored on GitHub's own container registry, ghcr.io.